Main version of Crawler (for Tentree client)
At the moment this crawler will get ALL required info from a PRODUCT page on Tentree, excluding images. I commented out the portions of the code that call the functions to run. Not sure how the webpage will work, I assume I may need to leave the function call, but have the URL passed be the one from the barebone web page?